@@ -4,50 +4,65 @@
   "scripts": {
     "test": "npm run lint",
     "precommit": "npm test",
-    "lint": "eslint webpack src",
-    "fix": "eslint --fix webpack src",
-    "build": "cross-env NODE_ENV=production webpack --progress --color --config ./webpack/webpack.config.prod.js",
-    "start": "cross-env NODE_ENV=development webpack-dev-server --color --watch --open --config ./webpack/webpack.config.dev.js"
+    "lint": "slint webpack src",
+    "fix": "slint webpack src --fix",
+    "build": "webpack --progress --color --config ./webpack/webpack.config.prod.",
+    "start": "webpack-dev-server --progress --color --watch --open --config ./webpack/webpack.config.dev."
   },
   "devDependencies": {
-    "autoprefixer": "^7.2.5",
+    "autoprefixer": "^8.1.0",
     "babel-core": "^6.26.0",
-    "babel-eslint": "^8.2.1",
-    "babel-loader": "^7.1.2",
+    "babel-eslint": "^8.2.2",
+    "babel-loader": "^7.1.4",
+    "babel-plugin-syntax-do-expressions": "^6.13.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.1",
-    "case-sensitive-paths-webpack-plugin": "^2.1.1",
-    "clean-webpack-plugin": "^0.1.17",
-    "copy-webpack-plugin": "^4.3.1",
-    "cross-env": "^5.1.3",
-    "css-loader": "^0.28.8",
-    "eslint": "^4.15.0",
-    "eslint-config-standard": "^11.0.0-beta.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
+    "babel-preset-react": "^6.24.1",
+    "case-sensitive-paths-webpack-plugin": "^2.1.2",
+    "clean-webpack-plugin": "^0.1.19",
+    "copy-webpack-plugin": "^4.5.0",
+    "css-loader": "^0.28.10",
+    "eslint": "^4.18.2",
+    "eslint-config-standard": "^11.0.0",
+    "eslint-config-standard-react": "^6.0.0",
+    "eslint-plugin-import": "^2.9.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-node": "^6.0.1",
     "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.7.0",
     "eslint-plugin-standard": "^3.0.1",
     "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.6",
+    "file-loader": "^1.1.11",
     "friendly-errors-webpack-plugin": "^1.6.1",
-    "html-webpack-plugin": "^2.30.1",
+    "html-webpack-plugin": "^3.0.6",
     "husky": "^0.14.3",
-    "less": "^2.7.3",
-    "less-loader": "^4.0.5",
-    "optimize-css-assets-webpack-plugin": "^3.2.0",
+    "less": "^3.0.1",
+    "less-loader": "^4.0.6",
+    "lodash.merge": "^4.6.1",
+    "optimize-css-assets-webpack-plugin": "^4.0.0",
     "postcss-less": "^1.1.3",
-    "postcss-loader": "^2.0.10",
-    "style-loader": "^0.19.1",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.11.0"
+    "postcss-loader": "^2.1.1",
+    "style-loader": "^0.20.2",
+    "ts-loader": "^4.0.1",
+    "ts-node": "^5.0.1",
+    "tslint": "^5.9.1",
+    "tslint-config-standard": "^7.0.0",
+    "tslint-react": "^3.5.1",
+    "typescript": "^2.7.2",
+    "url-loader": "^1.0.1",
+    "webpack": "^4.1.1",
+    "webpack-cli": "^2.0.10",
+    "webpack-dev-server": "^3.1.0"
   },
   "dependencies": {
-    "@types/node": "^9.3.0",
-    "@types/webpack": "^3.8.2"
+    "@types/node": "^9.4.6",
+    "@types/react": "^16.0.40",
+    "@types/react-dom": "^15.5.7",
+    "@types/webpack": "^4.0.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0"
   }
 }
