@@ -1,47 +1,48 @@
 {
   "name": "generator-fromzero",
-  "version": "0.2.0",
+  "version": "1.0.0",
   "description": "快速生成JavaScript开发环境",
-  "homepage": "http://mengb.net/coding.php/generator-fromzero",
+  "homepage": "https://github.com/ShirasawaSama/generator-fromzero.git",
   "author": "Shirasawa",
   "main": "index.js",
   "keywords": [
     "fromzero",
     "yeoman-generator",
     "webpack",
-    "eslint"
+    "eslint",
+    "babel",
+    "typescript",
+    "react",
+    "tslint"
   ],
   "files": [
     "index.js",
     "generators"
   ],
   "devDependencies": {
-    "babel-eslint": "^8.0.3",
-    "eslint": "^4.13.1",
-    "eslint-config-standard": "^11.0.0-beta.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
+    "babel-eslint": "^8.2.2",
+    "eslint": "^4.18.2",
+    "eslint-config-standard": "^11.0.0",
+    "eslint-plugin-import": "^2.9.0",
+    "eslint-plugin-node": "^6.0.1",
     "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "husky": "^0.14.3"
+    "eslint-plugin-standard": "^3.0.1"
   },
   "dependencies": {
-    "chalk": "^2.3.0",
+    "chalk": "^2.3.2",
     "generator-license": "^5.2.0",
-    "lodash": "^4.17.4",
+    "lodash": "^4.17.5",
     "mkdirp": "^0.5.1",
-    "yeoman-generator": "^2.0.1"
+    "yeoman-generator": "^2.0.3"
   },
   "scripts": {
     "test": "npm run lint",
-    "precommit": "npm test",
-    "prepublish": "node prepublish",
     "lint": "eslint index.js generators",
     "fix": "eslint --fix index.js generators"
   },
   "repository": {
     "type": "git",
-    "url": "http://mengb.net/coding.php/generator-fromzero"
+    "url": "git+https://github.com/ShirasawaSama/generator-fromzero.git"
   },
   "license": "MIT"
 }
